refactor(components): move normalizeMsg to prop schema
