+++ packages/components/src/functional-components/FormFieldMsg/FormFieldMsg.tsx
@@ -2,8 +2,7 @@ import { type FunctionalComponent, h } from '@stencil/core';
 import type { JSXBase } from '@stencil/core/internal';
 import clsx from 'clsx';
 
-import { type AlertPropType, type IdPropType, type MsgPropType, type Stringified } from '../../schema';
-import { normalizeMsg } from '../../utils/normalize-msg';
+import { type AlertPropType, type IdPropType, type MsgPropType, normalizeMsg, type Stringified } from '../../schema';
 import KolAlertFc from '../Alert';
 
 type FormFieldMsgProps = JSXBase.HTMLAttributes<HTMLDivElement> & {
